Renamed dinos.rb to dino.rb and updated require statements in the other files.
