+++ dino_catalog/dino.rb
