+++ dino_catalog/main.rb
@@ -1,5 +1,5 @@
 require 'csv'
-require_relative './dinos.rb'
+require_relative './dino.rb'
 require_relative './dino_filters.rb'
 require_relative './read_data.rb'
